requirements.txt
django==1.11.3
pytz==2017.2
django-appconf==1.0.2
django-compressor==2.1.1
django-axes==2.3.3
django-crispy-forms==1.6.1
django-timezone-field==2.0
django-braces==1.11.0
django-taggit==0.22.1